fix(security): client alias uniqueness fix(security): createClient, updateClient admin route return types
While not providing a security client alias, one gets autogenerated
based on the platform type, client id and, in case of web clients,
the domain name.

The return type fix for createClient and updateClient is technically a breaking change.
Not marking this as such to avoid bumping during rc stage.